Add files via upload
What This Project Does
Imagine you're planning a train journey, and you want to make sure you get a seat. This project helps you keep an eye on seat availability using a special tool – a Telegram Bot. The bot tells you when the seat count on your chosen train falls below a certain number you set. It's like having a friendly assistant that sends you a message whenever it's time to book your seat!

How It Works
Setting Things Up

First, you need to copy the project to your computer. It's like bringing a recipe into your kitchen.
You install the "ingredients" your recipe needs, which are some special tools (we call them "libraries").
You get some special keys – like keys to a secret door – from two different places: one for the train information and another for the Telegram Bot.
Running the Bot

Now, it's time to run the Telegram Bot. Just like turning on a TV.
The bot asks you for information about the train you want to track and how many seats you want to keep an eye on.
It's like telling your assistant which train you're interested in and when to warn you.
Checking Seats

The bot uses the "keys" you gave it to ask the internet about the train's seats.
It finds out how many seats are left and compares it to the number you want.
If the seats are getting low, it sends you a message on Telegram – like a friend sending a text to remind you of something.
Making It Your Own
Customizing the Bot

If you want, you can make the bot look and work the way you like. Just like decorating your room!
You can tell it how to talk to you and what messages to send.
You can even set up the bot to track different trains or seat numbers – it's all up to you.
Getting Help

Don't worry if you don't understand everything at first. You can ask the creator of the bot for help – just like asking a teacher if you don't understand a lesson.
They even left their email and a link to their place on the internet (GitHub) so you can talk to them and learn more.
Sharing the Recipe

This project is like a recipe for a magic dish. You can share it with your friends, and they can make their own bots too.
Remember, this project is like a way to learn about computer programs and how to make them do cool things!